boton borrar

+++ src/lib/view/templateMuro.js
@@ -6,6 +6,7 @@ export const muro = () => {
   const divMuro = document.createElement('div');
   const ViewMuro = `<div id="muro"> 
   <form id ="form_muro"><h3 class="titulo_muro">¿Qué estas pensando?</h3>
+  <p  class="titulo_muro">${displayNameData}</p>
   <textarea name="" id="coment_muro" cols="20" rows="10"></textarea>
   <button id="btn_muro">Publicar</button>
   </form>
@@ -37,6 +38,7 @@ export const muro = () => {
         console.error('Error adding document: ', error);
       });
   });
+
   const firestore = firebase.firestore();
   firestore.collection('coment', 'user').onSnapshot((querySnapshot) => {
     const lista = document.getElementById('public_muro');
@@ -50,28 +52,43 @@ export const muro = () => {
                
                  </div>
                 <button id="delete" value="${doc.id}">Borrar</button>
-                <button class="">Me gusta</button>
+                <button class="" onlick="editar('${doc.id}')">Editar</button>
                 </div>
                 <div class="commentDiv">
                   </div>`;
+      // conseguir id de coment
+      // const coment = firestore.collection('coment');
+      // coment
+      //   .onSnapshot((snap) => {
+      //     const usuario = [];
+      //     snap.forEach((snaphijo) => {
+      //       usuario.push({
+      //         id: snaphijo.id,
+      //         ...snaphijo.data(),
+      //       });
+      //     });
+      //     console.log(usuario);
+      //   });
 
-      // console.log(e.target.id);
-    });
-    const borrar = divMuro.querySelectorAll('#delete');
-    borrar.forEach((deletebutton) => {
-      deletebutton.addEventListener('click', (e) => {
-        console.log(e.target.value);
-        // console.log(doc.data().userid);
-        // console.log(uid);
-        firestore.collection('coment').doc(e.target.value).delete()
-          .then(() => {
-            if (uid === e.target.value) {
-              console.log('imprimir');
-            }
-          })
-          .catch((error) => {
-            console.error('Error removing document: ', error);
-          });
+      const borrar = divMuro.querySelectorAll('#delete');
+      borrar.forEach((deletebutton) => {
+        deletebutton.addEventListener('click', (e) => {
+          console.log(e.target.value);
+          console.log(uid);
+          console.log(doc.data().userid);
+          if (uid === doc.data().userid) {
+            alert('seguro deseas eliminar');
+            firestore.collection('coment').doc(e.target.value).delete()
+              .then()
+              .catch((error) => {
+                console.error('Error removing document: ', error);
+              });
+            console.log('borrado satisfactoriamente :)');
+          } else {
+            alert('No puedes elimiar un mensaje que no es tuyo');
+            console.log('son diferente id');
+          }
+        });
       });
     });
   });
